refactor(upload): seperate dropzone component
