+++ shared/design/src/components/upload/upload.tsx
@@ -4,11 +4,6 @@ import { Slot } from "radix-ui"
 import * as React from "react"
 
 import {
-  createClickHandler,
-  createDragHandlers,
-  createKeyboardHandler,
-} from "./core/dragAndDrop"
-import {
   createStore,
   StoreContext,
   useLazyRef,
@@ -28,7 +23,6 @@ import {
   type FileState,
   type FileUploadClearProps,
   type FileUploadContextValue,
-  type FileUploadDropzoneProps,
   type FileUploadItemContextValue,
   type FileUploadItemDeleteProps,
   type FileUploadItemMetadataProps,
@@ -41,7 +35,6 @@ import {
 
 const {
   ROOT: ROOT_NAME,
-  DROPZONE: DROPZONE_NAME,
   TRIGGER: TRIGGER_NAME,
   LIST: LIST_NAME,
   ITEM: ITEM_NAME,
@@ -58,11 +51,10 @@ function useDirection(dirProp?: Direction): Direction {
   return dirProp ?? contextDir ?? "ltr"
 }
 
-const FileUploadContext = React.createContext<FileUploadContextValue | null>(
-  null,
-)
+export const FileUploadContext =
+  React.createContext<FileUploadContextValue | null>(null)
 
-function useFileUploadContext(consumerName: string) {
+export function useFileUploadContext(consumerName: string) {
   const context = React.useContext(FileUploadContext)
   if (!context) {
     throw new Error(`\`${consumerName}\` must be used within \`${ROOT_NAME}\``)
@@ -261,96 +253,6 @@ function FileUploadRoot(props: FileUploadRootProps) {
         </RootPrimitive>
       </FileUploadContext.Provider>
     </StoreContext.Provider>
-  )
-}
-
-function FileUploadDropzone(props: FileUploadDropzoneProps) {
-  const {
-    asChild,
-    className,
-    onClick: onClickProp,
-    onDragOver: onDragOverProp,
-    onDragEnter: onDragEnterProp,
-    onDragLeave: onDragLeaveProp,
-    onDrop: onDropProp,
-    onPaste: onPasteProp,
-    onKeyDown: onKeyDownProp,
-    ...dropzoneProps
-  } = props
-
-  const context = useFileUploadContext(DROPZONE_NAME)
-  const store = useStoreContext(DROPZONE_NAME)
-  const dragOver = useStore((state) => state.dragOver)
-  const invalid = useStore((state) => state.invalid)
-
-  // 드래그 앤 드롭 핸들러들 생성
-  const dragHandlers = React.useMemo(
-    () =>
-      createDragHandlers({
-        store,
-        inputRef: context.inputRef,
-        existingHandlers: {
-          onDragOver: onDragOverProp,
-          onDragEnter: onDragEnterProp,
-          onDragLeave: onDragLeaveProp,
-          onDrop: onDropProp,
-          onPaste: onPasteProp,
-        },
-      }),
-    [
-      store,
-      context.inputRef,
-      onDragOverProp,
-      onDragEnterProp,
-      onDragLeaveProp,
-      onDropProp,
-      onPasteProp,
-    ],
-  )
-
-  // 클릭 핸들러 생성
-  const onClick = React.useCallback(
-    (event: React.MouseEvent<HTMLDivElement>) => {
-      const handler = createClickHandler(context.inputRef, onClickProp)
-      handler(event)
-    },
-    [context.inputRef, onClickProp],
-  )
-
-  // 키보드 핸들러 생성
-  const onKeyDown = React.useCallback(
-    (event: React.KeyboardEvent<HTMLDivElement>) => {
-      const handler = createKeyboardHandler(context.inputRef, onKeyDownProp)
-      handler(event)
-    },
-    [context.inputRef, onKeyDownProp],
-  )
-
-  const DropzonePrimitive = asChild ? Slot.Root : "div"
-
-  return (
-    <DropzonePrimitive
-      role="region"
-      id={context.dropzoneId}
-      aria-controls={`${context.inputId} ${context.listId}`}
-      aria-disabled={context.disabled}
-      aria-invalid={invalid}
-      data-disabled={context.disabled ? "" : undefined}
-      data-dragging={dragOver ? "" : undefined}
-      data-invalid={invalid ? "" : undefined}
-      data-slot="file-upload-dropzone"
-      dir={context.dir}
-      tabIndex={context.disabled ? undefined : 0}
-      {...dropzoneProps}
-      className={className}
-      onClick={onClick}
-      onDragEnter={dragHandlers.onDragEnter}
-      onDragLeave={dragHandlers.onDragLeave}
-      onDragOver={dragHandlers.onDragOver}
-      onDrop={dragHandlers.onDrop}
-      onKeyDown={onKeyDown}
-      onPaste={dragHandlers.onPaste}
-    />
   )
 }
 
@@ -742,10 +644,8 @@ function FileUploadClear(props: FileUploadClearProps) {
 
 export {
   FileUploadClear as Clear,
-  FileUploadDropzone as Dropzone,
   FileUploadRoot as FileUpload,
   FileUploadClear,
-  FileUploadDropzone,
   FileUploadItem,
   FileUploadItemDelete,
   FileUploadItemMetadata,
